+++ src/index.js
@@ -11,3 +11,4 @@ import './blocks/offices-block';
 import './blocks/agents-block';
 import './blocks/listing-block';
 import './blocks/leads-form-block';
+import './blocks/testimonials-block';